New serial command, merged WPS optional and dimmed display fixes
Added a serial command processor, that allows one to enter SSID and PASSWORD via the serial port.  Also added the "WPS is optional" feature as well as the fix for "should the display be dimmed", developed by Smitty-Halibut (I think?).  I also wrote a version of that last one, but his is prettier.